Can create hybrid levels from Random + Edited Supports Popup Messages now AI is faster now, but will ignore you if far away Performance boost via pruning draw-calls of ShadeLightCell's
New Focus rules - Accuracy, Evasion and Focus Fatigue (!)
Re-balanced Toughness and Rest Modes
HQ is back - can pick skills and perks - but UI is still glitched

+++ Dueling-Club/src/main/java/eidolons/libgdx/texture/TextureManager.java
@@ -159,7 +159,7 @@ public class TextureManager {
          getXY(origPath).getValue();
     }
 
-    public static float getFrameNumber(String path) {
+    public static int getFrameNumber(String path) {
         return getRows(path) * getColumns(path);
     }
